% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{go_paste}
\alias{go_paste}
\title{Send Code Copied to the Clipboard to Tidy Data Tutor}
\usage{
go_paste()
}
\value{
A string with an appropriately formatted URL to Tidy Data Tutor
(invisibly).
}
\description{
Send Code Copied to the Clipboard to Tidy Data Tutor
}
\examples{
\dontrun{

# Copy the following code into the clipboard:

library(dplyr)
library(palmerpenguins)

penguins \%>\%
  select(species, bill_length_mm) \%>\%
  group_by(species) \%>\%
  arrange(desc(bill_length_mm), .by_group = TRUE) \%>\%
  slice(1)

# Then run this in the console:

tidydatatutor::go_paste()
}
}
