% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{go_selection}
\alias{go_selection}
\title{Send Code Selected in the RStudio Editor to Tidy Data Tutor}
\usage{
go_selection()
}
\value{
A string with an appropriately formatted URL to Tidy Data Tutor
(invisibly).
}
\description{
Send Code Selected in the RStudio Editor to Tidy Data Tutor
}
\examples{
\dontrun{

# Select some pipeline code in the RStudio editor,
# then run this in the console:

tidydatatutor::go_selection()
}
}
