% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdt_url.R
\name{tdt_file}
\alias{tdt_file}
\title{Create a Link to Tidy Data Tutor from an R Code File}
\usage{
tdt_file(file)
}
\arguments{
\item{file}{The path to an R code file.}
}
\value{
A string with an appropriately formatted URL to Tidy Data Tutor.
}
\description{
Create a Link to Tidy Data Tutor from an R Code File
}
\examples{
\dontrun{

r_code_file <- system.file("test", "test-1.R", package = "tidydatatutor")
tdt_file(r_code_file)
}
}
