% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdt_url.R
\name{tdt_link}
\alias{tdt_link}
\title{Create a Link to Tidy Data Tutor}
\usage{
tdt_link(code)
}
\arguments{
\item{code}{A string or vector of strings containing R code.}
}
\value{
A string with an appropriately formatted URL to Tidy Data Tutor.
}
\description{
Create a Link to Tidy Data Tutor
}
\examples{
\dontrun{

code <- "library(dplyr)
Formaldehyde \%>\% mutate(Sum = carb + optden)"

tdt_link(code)
}
}
